Seprate generate component in the util
